Feature/crawling (#12)
* Add proxy URL for bypass CORS (#4)


* Revert "Add proxy URL for bypass CORS (#4)"

This reverts commit d78b5d16

* Add internet permission in AndroidManifest.xml

* 크롤링 기능 향상 (#9)

* Add proxy URL for bypass CORS

* Add Some useless file to gitignore

* 멜론 크롤링 기능 로직 향상, 폴더 이름 수정, 기존 searchLyric 함수 세분화

* Rollback .gitignore file

* Rollback pubspec.lock file

* Delete unuse import(meterial)

* Edit comment, Add TODO

* Enhance comments refer to https://dart.dev/guides/language/effective-dart/documentation

* Exclude further explanation about some variables

* Fix comments

* Remove duplicate comment
