+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.p_lyric">
 
+
+    <uses-permission android:name="android.permission.INTERNET"/>
+    
     <application
         android:icon="@mipmap/ic_launcher"
         android:label="p_lyric">
